+++ packages/web-components/src/pb-slot-context.ts
@@ -6,7 +6,8 @@ register(
   'pb-slot-context',
   () =>
     class PbSlotContext extends HTMLElement {
-      context: any = null;
+      _context: any = null;
+
       constructor() {
         super();
         this.attachShadow({ mode: 'open' });
@@ -17,15 +18,40 @@ register(
           // If the callback throws, propagation is already stopped
           event.callback(this.context || {});
         });
+      }
 
+      distributeContext() {
         const slots = Array.from(this.querySelectorAll('pb-slot')) as any[];
         const blocks = Array.from(this.querySelectorAll('pb-block')) as any[];
         for (const slot of slots) {
-          slot.context = this.context;
+          slot.context = this._context;
         }
         for (const block of blocks) {
-          block.context = this.context;
+          block.context = this._context;
         }
+      }
+
+      refreshContext() {
+        const name = this.getAttribute('context-name');
+        const value = this.getAttribute('context-value');
+        if (name) {
+          // Walk up parent elements and find the first one that has a context
+          const fullContext: any = { [name]: value };
+          let parent = this.parentElement;
+          while (parent) {
+            if ((parent as any).context) {
+              Object.assign(fullContext, (parent as any).context);
+            }
+            parent = parent.parentElement;
+          }
+
+          this._context = fullContext;
+        }
+      }
+
+      get context() {
+        this.refreshContext();
+        return this._context;
       }
 
       connectedCallback() {
@@ -34,17 +60,20 @@ register(
         const name = this.getAttribute('context-name');
         const value = this.getAttribute('context-value');
         if (name) {
-          this.context = Object.assign(this.context || {}, {
-            [name]: value,
-          });
+          // Walk up parent elements and find the first one that has a context
+          const fullContext: any = { [name]: value };
+          let parent = this.parentElement;
+          while (parent) {
+            if ((parent as any).context) {
+              Object.assign(fullContext, (parent as any).context);
+            }
+            parent = parent.parentElement;
+          }
+
+          this._context = fullContext;
         }
 
-        // Contexts above.
-        this.dispatchEvent(
-          new ContextEvent(pageContext, (newContext) => {
-            this.context = Object.assign(newContext || {}, this.context || {});
-          })
-        );
+        this.refreshContext();
 
         this.shadowRoot!.innerHTML = `<slot></slot>`;
       }