+++ config/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ui
 
+## 0.0.5
+
+### Patch Changes
+
+- @page-blocks/react@0.0.6
+
 ## 0.0.4
 
 ### Patch Changes